project-4/requirements.txt
torch
transformers
pillow
streamlit
numpy
tqdm
datasets